Rename UIComponent into Renderable
