+++ common/app/controllers/DevComponentController.scala
@@ -2,8 +2,8 @@ package controllers
 
 import model.ApplicationContext
 import play.api.mvc.{Action, Controller}
-import uiComponent.TestComponent
-import uiComponent.core.Renderer
+import rendering.TestComponent
+import rendering.core.Renderer
 
 class DevComponentController(renderer: Renderer)(implicit ac: ApplicationContext) extends Controller {
