+++ common/test/rendering/core/JavascriptRenderingTest.scala
@@ -1,4 +1,4 @@
-package uiComponent.core
+package rendering.core
 
 import java.io.FileNotFoundException
